Add require for cli/ui (#25)
This is related to https://github.com/Betterment/demo_mode/pull/23. I
was too hasty and forgot to actually `require 'cli/ui'` as outlined in
the
[cli-ui](https://github.com/Shopify/cli-ui?tab=readme-ov-file#installation)
installation instructions.

I tested this in a project using demo_mode by pointing it at this branch
of the gem locally and this resolved the same issue mentioned in
https://github.com/Betterment/demo_mode/pull/23 (which that PR didn't
resolve on its own).