Introducing persona metadata (#58)
This pull request introduces several changes to the `demo_mode`
functionality, primarily focusing on adding support for metadata in
sessions and ensuring proper handling of sessions in test environments.
The most important changes include the addition of metadata support,
adjustments to session handling in tests, and corresponding updates to
the test suite.

**Follow-up needed to clean up all the job adapter setting!**

### Metadata support:
*
[`app/models/demo_mode/session.rb`](diffhunk://#diff-0f5153bb33011b630b517794bf263bb46b0117d52ced272798cd460e2cf0714cR25-R36):
Added a `signinable_metadata` method to retrieve metadata when a session
is successful, and delegated the `metadata` method to the `persona`.
*
[`lib/demo_mode/persona.rb`](diffhunk://#diff-dcad6a8f33f3bdcc74ba2aecae17e21b00aa0ed4ce85c7fa406513ea01f828d9R41-R48):
Introduced a `metadata` method to store and retrieve metadata for
personas.
*
[`app/controllers/demo_mode/sessions_controller.rb`](diffhunk://#diff-7061eec77ccb3a1217893822312e65f7c45fde9996183101f9abfb0f3e7cc4a4R62):
Updated the `render_signinable_json` method to include metadata in the
response.

### Session handling in tests:
*
[`app/controllers/demo_mode/sessions_controller.rb`](diffhunk://#diff-7061eec77ccb3a1217893822312e65f7c45fde9996183101f9abfb0f3e7cc4a4R29):
Added a conditional reload of the session in the `create` method if the
queue adapter is set to 'test'.
*
[`spec/requests/demo_mode/sessions_spec.rb`](diffhunk://#diff-c5cfc8d30c2f238c6b2a4add285f4fb65c8cad4e12449097d780f7d1c5c3c4b1R140-R190):
Added tests to ensure sessions are created synchronously and metadata is
correctly returned when using the test adapter.
*
[`spec/system/demo_splash_spec.rb`](diffhunk://#diff-013651d6382adace2c19a1df70bdf5fe03771463cfecc54a0d8fc55d879c5ebdR7-R14):
Ensured the queue adapter is set to `inline` during tests to handle
sessions synchronously.