Update RSpec global job queue adapter to :test instead (#60)
This pull request includes several changes to the `demo_mode`
functionality, primarily focusing on the handling of the ActiveJob queue
adapter and related tests. The most important changes include
simplifying the queue adapter handling in the `SessionsController` and
updating the associated tests to reflect these changes.

Changes to queue adapter handling:

*
[`app/controllers/demo_mode/sessions_controller.rb`](diffhunk://#diff-7061eec77ccb3a1217893822312e65f7c45fde9996183101f9abfb0f3e7cc4a4L29-R29):
Removed the conditional check for the queue adapter name when reloading
the session. Now, the session reloads unconditionally.

Updates to test files:

*
[`spec/requests/demo_mode/sessions_spec.rb`](diffhunk://#diff-c5cfc8d30c2f238c6b2a4add285f4fb65c8cad4e12449097d780f7d1c5c3c4b1L106-R107):
Modified expectations for session creation tests to reflect changes in
session processing and user credentials. Also, changed the context from
'test adapter' to 'inline adapter' and updated the setup accordingly.
[[1]](diffhunk://#diff-c5cfc8d30c2f238c6b2a4add285f4fb65c8cad4e12449097d780f7d1c5c3c4b1L106-R107)
[[2]](diffhunk://#diff-c5cfc8d30c2f238c6b2a4add285f4fb65c8cad4e12449097d780f7d1c5c3c4b1L134-R141)
*
[`spec/spec_helper.rb`](diffhunk://#diff-89eebfcbc0f14b6d989517837ca1e94fce4e2ce9a03233641cd936f2b8d2ed94R48-R51):
Added a configuration to set the ActiveJob queue adapter to `:test`
before each test.
*
[`spec/system/demo_splash_spec.rb`](diffhunk://#diff-013651d6382adace2c19a1df70bdf5fe03771463cfecc54a0d8fc55d879c5ebdL7-R14):
Simplified the setup by directly setting the ActiveJob queue adapter to
`:inline` or `:async` in the `before` blocks instead of using `around`
blocks.
[[1]](diffhunk://#diff-013651d6382adace2c19a1df70bdf5fe03771463cfecc54a0d8fc55d879c5ebdL7-R14)
[[2]](diffhunk://#diff-013651d6382adace2c19a1df70bdf5fe03771463cfecc54a0d8fc55d879c5ebdL245-L250)